+++ bundler/src/bundler/chunk/mod.rs
@@ -15,6 +15,7 @@ mod export;
 mod merge;
 mod plan;
 mod remark;
+mod sort;
 
 #[derive(Debug)]
 struct InternalEntry {
